Add new test project and initial test files
A new project, CodeChallengeTests, has been added to the solution, with the intention of it hosting the unit tests for the application. This includes adding a csproj file with relevant package references and various sample test files. Unit tests are vital part of development in order to validate functionality and regressive issues.