Add Category model and update related models and migrations
This commit includes the addition of a new Category model. Changes are also made in related models, migrations, and DbContext to ensure proper foreign key relation and constraints with the Product model. Further, sample data for this new Category model is added to the exampleData.sql for ease of testing.