Implement nested set model for categories
This commit adds new migration files for applying the nested set model to categories. It includes the addition of 'Lft' and 'Rgt' columns to the 'Categories' table as well as the necessary navigation properties to build a tree-like structure. Empty migration files that were committed without changes have also been added.