Update data handling and logging level
Refactored the way data is reset in the database, updated the handling of 'CategoryId' in category and product models to optionally be null, and changed the default logging level from "Information" to "Warning". Furthermore, 'InsertController' has been renamed to 'ADataController'. Changes aimed at improving code clarity, reliability, and log efficiency.